requirements.txt
beautifulsoup4==4.12.3
fuzzywuzzy==0.18.0
librosa==0.10.2.post1
nest_asyncio==1.6.0
numpy==2.0.0
openai==1.34.0
pyperclip==1.8.2
python-dotenv==1.0.1
Requests==2.32.3
ruaccent==1.5.7.1
RUTTS==0.1.13
RUTTS==0.1.13
scipy==1.13.1
sounddevice==0.4.7
soundfile==0.12.1
SpeechRecognition==3.10.4
Telethon==1.36.0
urllib3==2.2.2
PyAudio==0.2.14
python-Levenshtein==0.25.1